requirements.txt
django >= 1.7.0
django-crispy-forms >= 1.4.0
python-social-auth >= 0.2.1
django-debug-toolbar >= 1.3.0